added noticle, change feed page
