+++ alumate/urls.py
@@ -11,4 +11,5 @@ urlpatterns = [
     path('chat/', include('chat.urls', namespace='chat')),
     path('inquiry/', include('inquiry.urls', namespace='inquiry')),
     path('people/', include('people.urls', namespace='people')),
+    path('noticle/', include('noticle.urls', namespace='noticle')),
 ]
