+++ inquiry/templates/inquiry/inquiry.html
@@ -3,19 +3,24 @@
 {% block navbar %}
 {% include 'include/nav-authpage.html' %}
 {% endblock %}
-{% block content %}
-  
+
+{% block style %}
 <style>
-    .card-zoom[data-animation=zooming] {
-        transition: all .3s cubic-bezier(.2,1,.22,1);
-        -webkit-transition: all .3s cubic-bezier(.2,1,.22,1);
-    }
-    .card-zoom[data-animation=zooming]:hover {
-    transform: scale(1.07);
-    -webkit-transform: scale(1.07);
-    z-index: 2;
-    }
+  .card-zoom[data-animation=zooming] {
+      transition: all .3s cubic-bezier(.2,1,.22,1);
+      -webkit-transition: all .3s cubic-bezier(.2,1,.22,1);
+  }
+  .card-zoom[data-animation=zooming]:hover {
+  transform: scale(1.07);
+  -webkit-transform: scale(1.07);
+  z-index: 2;
+  }
 </style>
+{% endblock %}
+  
+
+{% block content %}
+
     <div class="wrapper">     
       <div class="main">
         <div class="pt-lg mb-5 container">
@@ -26,72 +31,7 @@
               <p class="lead mt-1">Share the information and richen the collective knowledge to maximize the possibility that all people are able to reach the write information. Making a inquiry is the best way to reach your goals. Answering questions is the best way to support your mates.</p>
             </div>
           </div>
-          <div class="row">
-            <div class="ct-example">
-              
-              <div class="dropdown">
-                <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                  Country
-                </button>
-                <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                  <form action="">
-                    <div class="dropdown-item">
-                      <div class="custom-control custom-checkbox mb-3">
-                        <input class="custom-control-input" id="customCheck1" type="checkbox">
-                        <label class="custom-control-label" for="customCheck1">United Kingdom</label>
-                      </div>
-                    </div>
-                    <div class="dropdown-item">
-                      <div class="custom-control custom-checkbox mb-3">
-                        <input class="custom-control-input" id="customCheck2" type="checkbox">
-                        <label class="custom-control-label" for="customCheck2">United States</label>
-                      </div>
-                    </div>
-                    <div class="dropdown-item">
-                      <div class="custom-control custom-checkbox mb-3">
-                        <input class="custom-control-input" id="customCheck3" type="checkbox">
-                        <label class="custom-control-label" for="customCheck3">Australia</label>
-                      </div>
-                    </div>
-                    <div class="text-center">
-                      <button type="submit" class="btn btn-primary">Apply</button>
-                    </div>
-                  </form>
-                </div>
-              </div>
-              <div class="dropdown">
-                <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                  School
-                </button>
-                <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                  <form action="">
-                    <div class="dropdown-item">
-                      <div class="custom-control custom-checkbox mb-3">
-                        <input class="custom-control-input" id="customCheck4" type="checkbox">
-                        <label class="custom-control-label" for="customCheck4">Harvard University</label>
-                      </div>
-                    </div>
-                    <div class="dropdown-item">
-                      <div class="custom-control custom-checkbox mb-3">
-                        <input class="custom-control-input" id="customCheck5" type="checkbox">
-                        <label class="custom-control-label" for="customCheck5">Cambridge University</label>
-                      </div>
-                    </div>
-                    <div class="dropdown-item">
-                      <div class="custom-control custom-checkbox mb-3">
-                        <input class="custom-control-input" id="customCheck6" type="checkbox">
-                        <label class="custom-control-label" for="customCheck6">Tokyo University</label>
-                      </div>
-                    </div>
-                    <div class="text-center">
-                      <button type="submit" class="btn btn-primary">Apply</button>
-                    </div>
-                  </form>
-                </div>
-              </div>
-            </div>
-          </div>
-
+          {% include 'include/search_filter.html' %}
           <div class="row">
             <div class="col-lg-4 col-md-6">
               <div class="card card-blog card-zoom" data-header="pattern" data-animation="zooming">
