--- post/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import Post, Query
-# Register your models here.
-
-admin.site.register(Post)
-admin.site.register(Query)
