+++ templates/base-authpage.html
@@ -9,6 +9,9 @@
   <meta name="author" content="Creative Tim">
   <title>Alumate</title>
   {% include 'base/css.html' %}
+  
+  {% block style %}
+  {% endblock %}
 </head>
 <body>
   <header class="header-global">
