+++ templates/include/search_filter.html
@@ -1,0 +1,64 @@
+<div class="row">
+    <div class="ct-example">
+      <div class="dropdown">
+        <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+          Country
+        </button>
+        <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+          <form action="">
+            <div class="dropdown-item">
+              <div class="custom-control custom-checkbox mb-3">
+                <input class="custom-control-input" id="customCheck1" type="checkbox">
+                <label class="custom-control-label" for="customCheck1">United Kingdom</label>
+              </div>
+            </div>
+            <div class="dropdown-item">
+              <div class="custom-control custom-checkbox mb-3">
+                <input class="custom-control-input" id="customCheck2" type="checkbox">
+                <label class="custom-control-label" for="customCheck2">United States</label>
+              </div>
+            </div>
+            <div class="dropdown-item">
+              <div class="custom-control custom-checkbox mb-3">
+                <input class="custom-control-input" id="customCheck3" type="checkbox">
+                <label class="custom-control-label" for="customCheck3">Australia</label>
+              </div>
+            </div>
+            <div class="text-center">
+              <button type="submit" class="btn btn-primary">Apply</button>
+            </div>
+          </form>
+        </div>
+      </div>
+      <div class="dropdown">
+        <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+          School
+        </button>
+        <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+          <form action="">
+            <div class="dropdown-item">
+              <div class="custom-control custom-checkbox mb-3">
+                <input class="custom-control-input" id="customCheck4" type="checkbox">
+                <label class="custom-control-label" for="customCheck4">Harvard University</label>
+              </div>
+            </div>
+            <div class="dropdown-item">
+              <div class="custom-control custom-checkbox mb-3">
+                <input class="custom-control-input" id="customCheck5" type="checkbox">
+                <label class="custom-control-label" for="customCheck5">Cambridge University</label>
+              </div>
+            </div>
+            <div class="dropdown-item">
+              <div class="custom-control custom-checkbox mb-3">
+                <input class="custom-control-input" id="customCheck6" type="checkbox">
+                <label class="custom-control-label" for="customCheck6">Tokyo University</label>
+              </div>
+            </div>
+            <div class="text-center">
+              <button type="submit" class="btn btn-primary">Apply</button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+</div>
