--- user_profile/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class UserProfileConfig(AppConfig):
-    name = 'user_profile'
